to figure out the customers who are not a loan defaulter, sanction loan to gain profit and   build a relationship with customer for mutual benefit for a long period.
Individuals all around the world in some way depend on banks to lend them loans for various reasons to overcome their financial constraints and achieve some personal goals. Due to the ever-changing economy and ever-increasing competition in the financial world, the activity of taking a loan has become inevitable. Also, small scale to large scale banking firms depends on the activity of lending out loans to earn profits for managing their affairs and to function smoothly at times of financial constraints.
With the improving banking sector in recent times and the increasing trend of taking loans, a large population applies for bank loans. By the help of this dataset we can predict different situations such as loan default and get some laudatory insights that can benefit the banks.